requirements.txt
langgraph
langchain
langchain-google-genai
langchain-community
duckduckgo-search
pydantic
markdown
fpdf2
